Add support for panning the level with the mouse
Now the level can be panned with the mouse with "middle-button + drag"
or "Ctrl + left-button + drag". Works in Paste Preview Mode and WASD
Mode too.

Setting the cursor position with the mouse is also improved; now the
cursor follows the mouse pointer even if you move it outside of the
level's bounds.